Fix AMT export when starting from nested protein grid view
Problem was that we were ending up trying to filter by RowId using a Protein Group id on a peptide column. Fixed by setting up the filter before forcing the column list to the AMT expected columns.

Includes some cleanup refactoring, including collapsing two abstract classes into one, adding @Override annotations, and other simplifications that were possible after removing the legacy MS2 views earlier.
