+++ ms2/src/org/labkey/ms2/peptideview/AbstractMS2RunView.java
@@ -17,25 +17,40 @@
 package org.labkey.ms2.peptideview;
 
 import org.apache.log4j.Logger;
+import org.jetbrains.annotations.Nullable;
 import org.labkey.api.data.ButtonBar;
+import org.labkey.api.data.ColumnHeaderType;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DataRegion;
 import org.labkey.api.data.DisplayColumn;
+import org.labkey.api.data.ExcelWriter;
 import org.labkey.api.data.MenuButton;
+import org.labkey.api.data.NestableQueryView;
+import org.labkey.api.data.NestedRenderContext;
+import org.labkey.api.data.RenderContext;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.Sort;
+import org.labkey.api.data.TSVGridWriter;
+import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.TableSelector;
 import org.labkey.api.query.DetailsURL;
+import org.labkey.api.query.FieldKey;
+import org.labkey.api.query.QueryNestingOption;
+import org.labkey.api.query.QueryService;
+import org.labkey.api.query.QuerySettings;
+import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
 import org.labkey.api.util.Pair;
 import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.DataView;
+import org.labkey.api.view.DisplayElement;
 import org.labkey.api.view.GridView;
 import org.labkey.api.view.NavTree;
+import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.ViewContext;
-import org.labkey.api.view.WebPartView;
 import org.labkey.ms2.HydrophobicityColumn;
 import org.labkey.ms2.MS2Controller;
 import org.labkey.ms2.MS2ExportType;
@@ -44,12 +59,11 @@ import org.labkey.ms2.MS2Modification;
 import org.labkey.ms2.MS2Run;
 import org.labkey.ms2.MassType;
 import org.labkey.ms2.RunListException;
+import org.labkey.ms2.SpectrumIterator;
 import org.labkey.ms2.SpectrumRenderer;
 import org.labkey.ms2.protein.ProteinManager;
 import org.labkey.ms2.protein.tools.GoLoader;
 import org.labkey.ms2.protein.tools.ProteinDictionaryHelpers;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.Controller;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -57,6 +71,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 
@@ -64,7 +79,7 @@ import java.util.Map;
  * User: jeckels
  * Date: Feb 22, 2006
  */
-public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
+public abstract class AbstractMS2RunView
 {
     private static Logger _log = Logger.getLogger(AbstractMS2RunView.class);
 
@@ -83,20 +98,18 @@ public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
         _runs = runs;
     }
 
-    public WebPartType createGridView(MS2Controller.RunForm form)
+    public AbstractMS2QueryView createGridView(MS2Controller.RunForm form)
     {
         return createGridView(form.getExpanded(), false);
     }
 
-    public abstract WebPartType createGridView(boolean expanded, boolean forExport);
+    public abstract AbstractMS2QueryView createGridView(boolean expanded, boolean forExport);
 
     public abstract GridView getPeptideViewForProteinGrouping(String proteinGroupingId, String columns) throws SQLException;
 
     public abstract void addSQLSummaries(SimpleFilter peptideFilter, List<Pair<String, String>> sqlSummaries);
 
     public abstract SQLFragment getProteins(ActionURL queryUrl, MS2Run run, MS2Controller.ChartForm form);
-
-    public abstract Map<String, SimpleFilter> getFilter(ActionURL queryUrl, MS2Run run);
 
     public Container getContainer()
     {
@@ -112,14 +125,14 @@ public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
         return groupURL.toString() + "&proteinGroupingId=";
     }
     
-    protected ButtonBar createButtonBar(Class<? extends Controller> exportAllAction, Class<? extends Controller> exportSelectedAction, String whatWeAreSelecting, DataRegion dataRegion)
+    protected ButtonBar createButtonBar(String whatWeAreSelecting, DataRegion dataRegion)
     {
         ButtonBar result = new ButtonBar();
 
         List<MS2ExportType> exportFormats = getExportTypes();
 
         ActionURL exportUrl = _url.clone();
-        exportUrl.setAction(exportAllAction);
+        exportUrl.setAction(MS2Controller.ExportAllPeptidesAction.class);
         MenuButton exportAll = new MenuButton("Export All");
         for (MS2ExportType exportFormat : exportFormats)
         {
@@ -133,7 +146,7 @@ public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
         result.add(exportAll);
 
         MenuButton exportSelected = new MenuButton("Export Selected");
-        exportUrl.setAction(exportSelectedAction);
+        exportUrl.setAction(MS2Controller.ExportSelectedPeptidesAction.class);
         exportSelected.setRequiresSelection(true);
         for (MS2ExportType exportFormat : exportFormats)
         {
@@ -332,13 +345,139 @@ public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
         return _runs[0];
     }
 
-    public abstract ModelAndView exportToTSV(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows, List<String> headers) throws IOException;
+    public Map<String, SimpleFilter> getFilter(ActionURL queryUrl)
+    {
+        NestableQueryView queryView = createGridView(false, false);
+        RenderContext context = queryView.createDataView().getRenderContext();
+        TableInfo tinfo = queryView.createTable();
 
-    public abstract ModelAndView exportToAMT(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows) throws IOException;
+        Sort sort = new Sort();
+        return Collections.singletonMap("Filter", context.buildFilter(tinfo, Collections.emptyList(), queryUrl, queryView.getDataRegionName(), Table.ALL_ROWS, Table.NO_OFFSET, sort));
+    }
 
-    public abstract ModelAndView exportToExcel(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows) throws IOException;
+    public void exportToExcel(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows) throws IOException
+    {
+        createGridView(form.getExpanded(), true).exportToExcel(response, selectedRows);
+    }
 
-    public abstract void exportSpectra(MS2Controller.ExportForm form, ActionURL currentURL, SpectrumRenderer spectrumRenderer, List<String> exportRows) throws IOException, RunListException;
+
+    public void exportToTSV(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows, List<String> headers) throws IOException
+    {
+        AbstractMS2QueryView gridView = createGridView(form.getExpanded(), true);
+        gridView.createRowIdFragment(selectedRows);
+        gridView.exportToTSV(response, headers);
+    }
+
+    public void exportToAMT(MS2Controller.ExportForm form, HttpServletResponse response, List<String> selectedRows) throws IOException
+    {
+        AbstractMS2QueryView ms2QueryView = createGridView(form.getExpanded(), true);
+        ms2QueryView.createRowIdFragment(selectedRows);
+
+        List<FieldKey> keys = new ArrayList<>();
+        keys.add(FieldKey.fromParts("Fraction", "Run", "Run"));
+        keys.add(FieldKey.fromParts("Fraction", "Fraction"));
+        keys.add(FieldKey.fromParts("Mass"));
+        keys.add(FieldKey.fromParts("Scan"));
+        keys.add(FieldKey.fromParts("RetentionTime"));
+        keys.add(FieldKey.fromParts("H"));
+        keys.add(FieldKey.fromParts("PeptideProphet"));
+        keys.add(FieldKey.fromParts("Peptide"));
+        ms2QueryView.setOverrideColumns(keys);
+
+        ms2QueryView.exportToTSV(response, getAMTFileHeader());
+    }
+
+    public void exportSpectra(MS2Controller.ExportForm form, ActionURL currentURL, SpectrumRenderer spectrumRenderer, List<String> exportRows) throws IOException, RunListException
+    {
+        List<MS2Run> runs = form.validateRuns();
+
+        // Choose a different iterator based on whether this is a nested view that may include protein group criteria
+        NestableQueryView queryView = createGridView(form);
+        SQLFragment sql = generateSubSelect(queryView, currentURL, exportRows, FieldKey.fromParts("RowId")).second;
+        try (SpectrumIterator iter = new QueryResultSetSpectrumIterator(runs, sql))
+        {
+            spectrumRenderer.render(iter);
+            spectrumRenderer.close();
+        }
+    }
+
+    /** Generate the SELECT SQL to get a particular FieldKey, respecting the filters and other config on the URL */
+    protected Pair<ColumnInfo, SQLFragment> generateSubSelect(NestableQueryView queryView, ActionURL currentURL, @Nullable List<String> selectedIds, FieldKey desiredFK)
+    {
+        RenderContext context = queryView.createDataView().getRenderContext();
+        TableInfo tinfo = queryView.createTable();
+
+        Sort sort = new Sort();
+        SimpleFilter filter;
+        if (context instanceof NestedRenderContext)
+        {
+            filter = ((NestedRenderContext)context).buildFilter(tinfo, Collections.emptyList(), currentURL, queryView.getDataRegionName(), Table.ALL_ROWS, Table.NO_OFFSET, sort, true);
+        }
+        else
+        {
+            filter = context.buildFilter(tinfo, Collections.emptyList(), currentURL, queryView.getDataRegionName(), Table.ALL_ROWS, Table.NO_OFFSET, sort);
+        }
+        addSelectionFilter(selectedIds, queryView, filter);
+
+        ColumnInfo desiredCol = QueryService.get().getColumns(tinfo, Collections.singletonList(desiredFK)).get(desiredFK);
+        if (desiredCol == null)
+        {
+            throw new IllegalArgumentException("Couldn't find column " + desiredFK + " in table " + tinfo);
+        }
+
+        List<ColumnInfo> columns = new ArrayList<>();
+        columns.add(desiredCol);
+
+        QueryService.get().ensureRequiredColumns(tinfo, columns, filter, sort, new HashSet<>());
+
+        SQLFragment sql = QueryService.get().getSelectSQL(tinfo, columns, filter, sort, Table.ALL_ROWS, Table.NO_OFFSET, false);
+        return new Pair<>(desiredCol, sql);
+    }
+
+    /** Add a filter for any selection the user might have made. The type of selection depends on the type of view (peptides/protein groups/search engine protein) */
+    private void addSelectionFilter(@Nullable List<String> exportRows, NestableQueryView queryView, SimpleFilter filter)
+    {
+        if (exportRows != null)
+        {
+            List<Integer> rowIds = parseIds(exportRows);
+            FieldKey selectionFK;
+            QueryNestingOption nestingOption = queryView.getSelectedNestingOption();
+            if (nestingOption != null)
+            {
+                // We're nested, so the selection key is going to be at the protein or protein group level
+                selectionFK = nestingOption.getAggregateRowIdFieldKey();
+            }
+            else
+            {
+                // No nesting, so the selection key will just be the peptide's RowId
+                selectionFK = FieldKey.fromParts("RowId");
+            }
+            filter.addClause(new SimpleFilter.InClause(selectionFK, rowIds));
+        }
+    }
+
+    /**
+     * Convert from Strings to Integers
+     * @throws NotFoundException if there's an unparseable value
+     */
+    private List<Integer> parseIds(List<String> exportRows)
+    {
+        List<Integer> rowIds = new ArrayList<>(exportRows.size());
+        for (String exportRow : exportRows)
+        {
+            try
+            {
+                rowIds.add(Integer.parseInt(exportRow));
+            }
+            catch (NumberFormatException e)
+            {
+                throw new NotFoundException("Invalid selection: " + exportRow);
+            }
+        }
+        return rowIds;
+    }
+
+
 
     protected List<String> getAMTFileHeader()
     {
@@ -370,5 +509,81 @@ public abstract class AbstractMS2RunView<WebPartType extends WebPartView>
         }
 
         return fileHeader;
+    }
+
+    public abstract class AbstractMS2QueryView extends NestableQueryView
+    {
+        protected SimpleFilter.FilterClause _selectedRowsClause;
+
+        public AbstractMS2QueryView(UserSchema schema, QuerySettings settings, boolean expanded, boolean forExport, QueryNestingOption... queryNestingOptions)
+        {
+            super(schema, settings, expanded, forExport, queryNestingOptions);
+
+            setViewItemFilter((type, label) -> SingleMS2RunRReport.TYPE.equals(type));
+        }
+
+        @Override
+        protected void populateButtonBar(DataView view, ButtonBar bar)
+        {
+            super.populateButtonBar(view, bar);
+            ButtonBar bb = createButtonBar("peptides", view.getDataRegion());
+            for (DisplayElement element : bb.getList())
+            {
+                bar.add(element);
+            }
+        }
+
+        public void exportToTSV(HttpServletResponse response, List<String> headers) throws IOException
+        {
+            getSettings().setMaxRows(Table.ALL_ROWS);
+
+            try (TSVGridWriter tsvWriter = getTsvWriter())
+            {
+                tsvWriter.setColumnHeaderType(ColumnHeaderType.Caption);
+                tsvWriter.setFileHeader(headers);
+                tsvWriter.write(response);
+            }
+        }
+
+        public void exportToExcel(HttpServletResponse response, List<String> selectedRows) throws IOException
+        {
+            createRowIdFragment(selectedRows);
+            getSettings().setMaxRows(ExcelWriter.MAX_ROWS_EXCEL_97);
+            exportToExcel(response);
+        }
+
+        protected void createRowIdFragment(List<String> selectedRows)
+        {
+            if (selectedRows != null)
+            {
+                List<Integer> parsedSelection = new ArrayList<>();
+                for (String selectedRow : selectedRows)
+                {
+                    Integer row = Integer.valueOf(selectedRow);
+                    parsedSelection.add(row);
+                }
+
+                // Don't used _selectedNestingOption one because we want to export as if we're a simple flat view
+                QueryNestingOption nesting = determineNestingOption();
+                FieldKey column = nesting == null ? FieldKey.fromParts("RowId") : nesting.getRowIdFieldKey();
+                _selectedRowsClause = new SimpleFilter.InClause(column, parsedSelection);
+            }
+        }
+
+        @Override
+        public DataView createDataView()
+        {
+            DataView result = super.createDataView();
+            SimpleFilter filter = new SimpleFilter(result.getRenderContext().getBaseFilter());
+
+            if (_selectedRowsClause != null)
+            {
+                filter.addClause(_selectedRowsClause);
+            }
+
+            filter.addAllClauses(ProteinManager.getPeptideFilter(_url, ProteinManager.EXTRA_FILTER | ProteinManager.PROTEIN_FILTER, getUser(), _runs));
+            result.getRenderContext().setBaseFilter(filter);
+            return result;
+        }
     }
 }
